+++ lua/onedark/extra/init.lua
@@ -1,7 +1,5 @@
 package.path = "./lua/?/init.lua;./lua/?.lua"
 
-local config_module = require("onedark.config")
-
 local function write(str, fileName)
   print("[write] extra/" .. fileName)
   local file = io.open("extras/" .. fileName, "w")
@@ -11,6 +9,7 @@ end
 
 local extras = {kitty = "conf", alacritty = "yml"}
 for extra, ext in pairs(extras) do
+  local config = {transform_colors = true}
   local plugin = require("onedark.extra." .. extra)
-  write(plugin[extra](config_module.config), extra .. "_onedark" .. "." .. ext)
+  write(plugin[extra](config), extra .. "_onedark" .. "." .. ext)
 end